requirements.txt
asgiref==3.5.2
Django==4.1.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.4.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.3.0
gunicorn==20.1.0